+++ cli/cmdmanager/manager_test.go
@@ -754,7 +754,27 @@ func TestCommandManager_ExecCommand(t *testing.T) {
 				m.Return(executionResponse, nil)
 			},
 			executionRequestFunc: func(c context.Context, execution command.ExecutionRequest) {
-				assert.Equal(t, "entrypoint arg1 arg2", execution.Command)
+				assert.Equal(t, "entrypoint \"arg1\" \"arg2\"", execution.Command)
+			},
+		},
+		{
+			name:  "executor success with quoted args",
+			alias: alias,
+			args:  []string{"\"arg1\"", "\"arg2\""},
+			assetGetterFunc: func(m *mockassetgetter.MockAssetGetter) {
+				runtimeAsset := asset.RuntimeAsset{
+					Path:   assetPath,
+					SHA512: checksum,
+				}
+				m.On("Get", ctx, &testAsset).
+					Return(&runtimeAsset, nil)
+			},
+			executorFunc: func(m *mockexecutor.MockExecutor) {
+				executionResponse := command.FixtureExecutionResponse(0, "success\n")
+				m.Return(executionResponse, nil)
+			},
+			executionRequestFunc: func(c context.Context, execution command.ExecutionRequest) {
+				assert.Equal(t, "entrypoint \"\\\"arg1\\\"\" \"\\\"arg2\\\"\"", execution.Command)
 			},
 		},
 	}