+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Errors while publishing proxy check requests do not block scheduling for other
 entities.
 - Listing namespaces in sensuctl with `--chunk-size` now works properly.
+- Escaped quotes are no longer required for asset arguments run by `sensuctl command exec`.
 
 ## [5.20.2] - 2020-05-26